requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
aiohttp==3.8.6
asyncio==3.4.3
websockets==11.0.3
python-multipart==0.0.6
email-validator==2.0.0
pyjwt==2.8.0
bcrypt==4.0.1
passlib==1.7.4

# Database and caching
sqlalchemy==2.0.22
alembic==1.12.0
asyncpg==0.28.0
redis==5.0.1
motor==3.3.1
pymongo==4.5.0

# Vector databases
pinecone-client==2.2.4
weaviate-client==3.24.1

# AI and ML
openai==1.3.5
langchain==0.0.335
langchain-openai==0.0.2.post1
langchain-community==0.0.9
crewai==0.28.5
sentence-transformers==2.2.2
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
transformers==4.34.0
torch==2.1.0
tiktoken==0.5.1

# Integrations
hubspot-api-client==8.2.1
simple-salesforce==1.12.4
gong-client==0.1.0
slack-sdk==3.23.0
snowflake-connector-python==3.5.0
airbyte-cdk==0.51.0

# Infrastructure
pulumi==3.91.1
pulumi-aws==6.5.0
pulumi-docker==4.4.1
pulumi-kubernetes==4.5.4
pulumi-random==4.14.0
pulumi-command==0.9.1

# Monitoring and logging
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-prometheus==1.20.0
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
pytest-mock==3.12.0

# Development
black==23.9.1
isort==5.12.0
mypy==1.6.1
flake8==6.1.0
pre-commit==3.5.0
